% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_quality_control.R
\name{adt_quality_control_defaults}
\alias{adt_quality_control_defaults}
\alias{computeAdtQcMetricsDefaults}
\alias{suggestAdtQcThresholdsDefaults}
\title{Default parameters for ADT quality control}
\usage{
computeAdtQcMetricsDefaults()

suggestAdtQcThresholdsDefaults()
}
\value{
Named list containing default values for the various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in each function's signature.
}
\examples{
computeAdtQcMetricsDefaults()
suggestAdtQcThresholdsDefaults()
}
\author{
Aaron Lun
}
