% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateAcrossCells.R
\name{aggregateAcrossCellsDefaults}
\alias{aggregateAcrossCellsDefaults}
\title{Default parameters for \code{\link{aggregateAcrossCells}}}
\usage{
aggregateAcrossCellsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{aggregateAcrossCells}} function signature.
}
\examples{
aggregateAcrossCellsDefaults()
}
\author{
Aaron Lun
}
