% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateAcrossGenes.R
\name{aggregateAcrossGenesDefaults}
\alias{aggregateAcrossGenesDefaults}
\title{Default parameters for \code{\link{aggregateAcrossGenes}}}
\usage{
aggregateAcrossGenesDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{aggregateAcrossGenes}} function signature.
}
\examples{
aggregateAcrossGenesDefaults()
}
\author{
Aaron Lun
}
