% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerSizeFactors.R
\name{centerSizeFactorsDefaults}
\alias{centerSizeFactorsDefaults}
\title{Default parameters for \code{\link{centerSizeFactors}}}
\usage{
centerSizeFactorsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters for \code{\link{centerSizeFactors}}
}
\examples{
centerSizeFactorsDefaults()
}
\author{
Aaron Lun
}
