% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerSpikeInFactors.R
\name{centerSpikeInFactors}
\alias{centerSpikeInFactors}
\title{Center spike-in size factors}
\usage{
centerSpikeInFactors(endogenous, spike.ins, block = NULL, mode = NULL)
}
\arguments{
\item{endogenous}{Numeric vector of size factors for endogenous genes across cells.
Invalid size factors (e.g., non-positive, non-finite) will be ignored.}

\item{spike.ins}{List of numeric vectors.
Each vector should be of length equal to \code{endogenous} and contains size factors for a set of spike-in transcripts.
Invalid size factors (e.g., non-positive, non-finite) will be ignored.}

\item{block}{Vector or factor of length equal to \code{endogenous}, specifying the block of origin for each cell.
Alternatively \code{NULL}, in which case all cells are assumed to be in the same block.}

\item{mode}{String specifying how to scale size factors across blocks, see the argument of the same name in \code{\link{centerSizeFactors}}.

If \code{NULL}, the default value in \code{\link{centerSpikeInFactors}} is used.

This argument is only used if \code{block} is provided.}
}
\value{
List containing:
\itemize{
\item \code{endogenous}, a numeric vector containing the centered size factors for the endogenous genes.
\item \code{spike.ins}, a named list of numeric vectors.
Each vector is named after an entry of \code{spike.ins} and contains centered size factors for the corresponding spike-in transcripts.
}
}
\description{
Center size factors for endogenous genes and spike-in transcripts, following the rationale in \code{\link{centerSizeFactors}}.
This aims to preserve the scale of the counts in each feature set so that the normalized expression values are still comparable.
}
\details{
This function is effectively a convenient wrapper around \code{\link{centerSizeFactors}},
configured to ensure that the mean of each set of spike-in size factors is the same as that of the endogenous size factors.
The aim is to preserve the scale of the average abundances of the spike-ins relative to the endogenous genes for a valid comparison between feature sets.
Typically, we would use spike-in transcripts to infer some properties of endogenous genes of similar molarity, e.g., technical variance.
}
\examples{
endogenous <- runif(100)
spike.ercc <- runif(100)
centerSpikeInFactors(endogenous, list(ERCC = spike.ercc))

block <- sample(3, 100, replace=TRUE)
centerSpikeInFactors(endogenous, list(ERCC = spike.ercc), block=block)

}
\seealso{
The \code{center_spike_in_factors} and \code{center_spike_in_factors_blocked} functions in \url{https://libscran.github.io/scran_norm/}.
}
\author{
Aaron Lun
}
