% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerSpikeInFactors.R
\name{centerSpikeInFactorsDefaults}
\alias{centerSpikeInFactorsDefaults}
\title{Default parameters for \code{\link{centerSpikeInFactors}}}
\usage{
centerSpikeInFactorsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters for \code{\link{centerSpikeInFactors}}
}
\examples{
centerSpikeInFactorsDefaults()
}
\author{
Aaron Lun
}
