% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseHighlyVariableGenes.R
\name{chooseHighlyVariableGenesDefaults}
\alias{chooseHighlyVariableGenesDefaults}
\title{Default parameters for \code{\link{chooseHighlyVariableGenes}}}
\usage{
chooseHighlyVariableGenesDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters for \code{\link{chooseHighlyVariableGenes}}
}
\examples{
chooseHighlyVariableGenesDefaults()
}
\author{
Aaron Lun
}
