% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePseudoCount.R
\name{choosePseudoCount}
\alias{choosePseudoCount}
\title{Choose a suitable pseudo-count}
\usage{
choosePseudoCount(size.factors, quantile = 0.05, max.bias = 1, min.value = 1)
}
\arguments{
\item{size.factors}{Numeric vector of size factors for all cells.}

\item{quantile}{Numeric scalar specifying the quantile to use for defining extreme size factors.}

\item{max.bias}{Numeric scalar specifying the maximum allowed bias.}

\item{min.value}{Numeric scalar specifying the minimum value for the pseudo-count.}
}
\value{
A choice of pseudo-count for \code{\link{normalizeCounts}}.
}
\description{
Choose a suitable pseudo-count to control the bias introduced by log-transformation of normalized counts.
}
\examples{
sf <- runif(100)
choosePseudoCount(sf)
choosePseudoCount(sf, quantile=0.01)
choosePseudoCount(sf, max.bias=0.5)

}
\seealso{
\url{https://libscran.github.io/scran_norm/}, for the motivation behind calculating a larger pseudo-count.
}
\author{
Aaron Lun
}
