% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKmeans.R
\name{clusterKmeansDefaults}
\alias{clusterKmeansDefaults}
\title{Default parameters for \code{\link{clusterKmeans}}}
\usage{
clusterKmeansDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters for \code{\link{clusterKmeans}}
}
\examples{
clusterKmeansDefaults()
}
\author{
Aaron Lun
}
