% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBlockWeights.R
\name{computeBlockWeightsDefaults}
\alias{computeBlockWeightsDefaults}
\title{Default parameters for \code{\link{computeBlockWeights}}}
\usage{
computeBlockWeightsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters for \code{\link{computeBlockWeights}}
}
\examples{
computeBlockWeightsDefaults()
}
\author{
Aaron Lun
}
