% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeClrm1Factors.R
\name{computeClrm1Factors}
\alias{computeClrm1Factors}
\alias{computeClrm1FactorsDefaults}
\title{Compute size factors for ADT counts}
\usage{
computeClrm1Factors(x, num.threads = NULL)

computeClrm1FactorsDefaults()
}
\arguments{
\item{x}{A matrix-like object containing ADT count data.
Rows correspond to tags and columns correspond to cells.}

\item{num.threads}{Number of threads to use.
If \code{NULL}, the default value in \code{computeClrm1FactorDefaults} is used.}
}
\value{
\code{computeClrm1Factors} returns a numeric vector containing the CLRm1 size factor for each cell.
Note that these size factors are not centered and should be passed through, e.g., \code{\link{centerSizeFactors}} before normalization.

\code{computeClrm1FactorsDefaults} returns a named list containing default values for the various function arguments.
}
\description{
Compute size factors from an ADT count matrix using the CLRm1 method.
This is a variant of the centered log-ratio (CLR) method, where the size factors are defined from the geometric mean of counts within each cell.
}
\examples{
library(Matrix)
x <- abs(rsparsematrix(1000, 100, 0.1) * 10)
head(computeClrm1Factors(x))

}
\seealso{
\url{https://github.com/libscran/clrm1}, for a description of the CLRm1 method.

\code{\link{normalizeAdtCounts.se}}, which computes CLRm1 factors prior to normalization.
}
\author{
Aaron Lun
}
