% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeClrm1Factors.R
\name{computeClrm1FactorsDefaults}
\alias{computeClrm1FactorsDefaults}
\title{Default parameters for \code{\link{computeClrm1Factors}}}
\usage{
computeClrm1FactorsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{computeClrm1Factors}} function signature.
}
\examples{
computeClrm1FactorsDefaults()
}
\author{
Aaron Lun
}
