% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctMnn.R
\name{correctMnnDefaults}
\alias{correctMnnDefaults}
\title{Default parameters for \code{\link{correctMnn}}}
\usage{
correctMnnDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{correctMnn}} function signature.
}
\examples{
correctMnnDefaults()
}
\author{
Aaron Lun
}
