% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_quality_control.R
\name{crispr_quality_control_defaults}
\alias{crispr_quality_control_defaults}
\alias{computeCrisprQcMetricsDefaults}
\alias{suggestCrisprQcThresholdsDefaults}
\title{Default parameters for CRISPR quality control}
\usage{
computeCrisprQcMetricsDefaults()

suggestCrisprQcThresholdsDefaults()
}
\value{
Named list containing default values for the various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in each function's signature.
}
\examples{
computeCrisprQcMetricsDefaults()
suggestCrisprQcThresholdsDefaults()
}
\author{
Aaron Lun
}
