% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{analyze}
\alias{analyze}
\alias{convertAnalyzeResults}
\title{Defunct functions}
\usage{
analyze(...)

convertAnalyzeResults(...)
}
\arguments{
\item{...}{Arguments, ignored.}
}
\description{
See Details for each function's recommended replacements.
}
\details{
\code{analyze} is replaced by \code{\link{analyze.se}}.

\code{convertAnalyzeResults} is no longer necessary as conversion is done within \code{\link{analyze.se}}.
}
