% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVarianceTrend.R
\name{fitVarianceTrendDefaults}
\alias{fitVarianceTrendDefaults}
\title{Default parameters for \code{\link{fitVarianceTrend}}}
\usage{
fitVarianceTrendDefaults()
}
\value{
Named list containing default values for the various function arguments.
}
\description{
Default parameters for \code{\link{fitVarianceTrend}}
}
\examples{
fitVarianceTrendDefaults()
}
\author{
Aaron Lun
}
