% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneVariances.R
\name{modelGeneVariances}
\alias{modelGeneVariances}
\title{Model per-gene variances in expression}
\usage{
modelGeneVariances(
  x,
  block = NULL,
  block.weight.policy = c("variable", "equal", "none"),
  variable.block.weight = c(0, 1000),
  mean.filter = TRUE,
  min.mean = 0.1,
  transform = TRUE,
  span = 0.3,
  use.min.width = FALSE,
  min.width = 1,
  min.window.count = 200,
  num.threads = 1
)
}
\arguments{
\item{x}{A matrix-like object where rows correspond to genes or genomic features and columns correspond to cells.
It is typically expected to contain log-expression values, e.g., from \code{\link{normalizeCounts}}.}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for each cell in \code{x}.
Alternatively \code{NULL} if all cells are from the same block.}

\item{block.weight.policy}{String specifying the policy to use for weighting different blocks when computing the average for each statistic
Only used if \code{block} is not \code{NULL}.}

\item{variable.block.weight}{Numeric vector of length 2, specifying the parameters for variable block weighting.
The first and second values are used as the lower and upper bounds, respectively, for the variable weight calculation.
Only used if \code{block} is not \code{NULL} and \code{block.weight.policy = "variable"}.}

\item{mean.filter}{Logical scalar indicating whether to filter on the means before trend fitting.}

\item{min.mean}{Numeric scalar specifying the minimum mean of genes to use in trend fitting.
Only used if \code{mean.filter=TRUE}.}

\item{transform}{Logical scalar indicating whether a quarter-root transformation should be applied before trend fitting.}

\item{span}{Numeric scalar specifying the span of the LOWESS smoother.
Ignored if \code{use.min.width=TRUE}.}

\item{use.min.width}{Logical scalar indicating whether a minimum width constraint should be applied to the LOWESS smoother.
Useful to avoid overfitting in high-density intervals.}

\item{min.width}{Minimum width of the window to use when \code{use.min.width=TRUE}.}

\item{min.window.count}{Minimum number of observations in each window.
Only used if \code{use.min.width=TRUE}.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing \code{statistics}.
This is a data frame with the columns \code{means}, \code{variances}, \code{fitted} and \code{residuals},
each of which is a numeric vector containing the statistic of the same name across all genes.

If \code{block} is supplied, each of the column vectors described above contains the average across all blocks.
The list will also contain \code{per.block}, a list of data frames containing the equivalent statistics for each block.
}
\description{
Compute the variance in (log-)expression values for each gene,
and model the trend in the variances with respect to the mean.
}
\examples{
library(Matrix)
x <- abs(rsparsematrix(1000, 100, 0.1) * 10)
out <- modelGeneVariances(x)
str(out)

# Throwing in some blocking.
block <- sample(letters[1:4], ncol(x), replace=TRUE)
out <- modelGeneVariances(x, block=block)
str(out)

}
\seealso{
The \code{model_gene_variances} function in \url{https://libscran.github.io/scran_variances/}, for the underlying implementation.

\code{\link{fitVarianceTrend}}, which fits the mean-variance trend.
}
\author{
Aaron Lun
}
