% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneVariances.R
\name{modelGeneVariancesDefaults}
\alias{modelGeneVariancesDefaults}
\title{Default parameters for \code{\link{modelGeneVariances}}}
\usage{
modelGeneVariancesDefaults()
}
\value{
Named list of default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{modelGeneVariances}} function signature.
}
\examples{
modelGeneVariancesDefaults()
}
\author{
Aaron Lun
}
