% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCountsDefaults}
\alias{normalizeCountsDefaults}
\title{Default parameters for \code{\link{normalizeCounts}}}
\usage{
normalizeCountsDefaults()
}
\value{
Named list of default values for various function arguments.
}
\description{
Default parameters for \code{\link{normalizeCounts}}
}
\examples{
normalizeCountsDefaults()
}
\author{
Aaron Lun
}
