% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_normalizeRnaCountsWithSpikeIns.R
\name{normalizeRnaCountsWithSpikeIns.se}
\alias{normalizeRnaCountsWithSpikeIns.se}
\title{Normalize RNA and spike-in counts}
\usage{
normalizeRnaCountsWithSpikeIns.se(
  x,
  spike.altexps,
  endogenous.factors = NULL,
  spike.factors = NULL,
  use.spike.ins.for.endogenous = FALSE,
  block = NULL,
  mode = "lowest",
  log = TRUE,
  pseudo.count = 1,
  more.norm.args = list(),
  assay.type = "counts",
  output.name = "logcounts",
  factor.name = "sizeFactor",
  num.threads = 1
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment]{SingleCellExperiment} or one of its subclasses.
The main experiment should contain count data for endogenous genes, where rows correspond to genes and columns correspond to cells.
There should also be one or more alternative experiments containing spike-in data, see \code{spike.altexps}.}

\item{spike.altexps}{Alternative experiments containing spike-in data.
This should be an unnamed integer or character vector containing the names/indices of the alternative experiments of interest.
The assay to use from each alternative experiment is determined by \code{assay.type}.

Alternatively, \code{spike.altexps} may be a named integer or character vector.
Each name specifies an alternative experiment while each value is the index/name of the assay to use from that experiment.}

\item{endogenous.factors}{Numeric vector of length equal to \code{ncol(x)}, containing the size factors for the endogenous genes.
If \code{NULL}, this defaults to the column sums of the relevant count matrix in \code{x}.
Ignored if \code{use.spike.ins.for.endogenous.factors} is not \code{FALSE}.}

\item{spike.factors}{Named list of numeric vectors containing the size factors for each spike-in set.
Each entry corresponds to a spike-in set ad be named after its alternative experiment in \code{spike.altexps}.
Each vector should be of length equal to \code{ncol(x)}. 
If \code{NULL} or a spike-in set is not present in \code{spike.factors}, the column sums of the relevant count matrix is used instead.}

\item{use.spike.ins.for.endogenous}{Boolean indicating whether to use spike-in factors for normalization of endogenous genes.
By default, each set of spike-in factors is only used on its corresponding count matrix, while the endogenous factors are used on the endogenous count matrix.

If this is set to \code{TRUE}, the size factors from the first spike-in set in \code{spike.altexps} are used to normalize the endogenous counts.
This is useful for removing technical differences in scaling while preserving differences in total RNA content between cells.

Alternatively, this may be a string specifying the name of the spike-in set from which to obtain size factors for normalization of the endogenous counts.
This should refer to one of the alternative experiments in \code{spike.altexps}.}

\item{block}{Block assignments for each cell, passed to \code{\link{centerSizeFactors}}.}

\item{mode}{How to center size factors in different blocks, see \code{?\link{centerSizeFactors}} for more details.}

\item{log}{Whether to log-transform the normalized expression values, see \code{?\link{normalizeCounts}} for more details.}

\item{pseudo.count}{The pseudo-count for log-transformation, see \code{?\link{normalizeCounts}} for more details.}

\item{more.norm.args}{Named list of additional arguments to pass to \code{\link{normalizeCounts}}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} with the count matrix.}

\item{output.name}{String containing the name of the assay to store the normalized matrix.}

\item{factor.name}{String containing the name of the \code{\link[SummarizedExperiment]{colData}} column in which to store the size factors in the output object.
If \code{NULL}, the size factors are not stored.}

\item{num.threads}{Integer specifying the number of threads for computing column sums.
Only used if precomputed factors are not available in \code{endogenous.factors} or \code{spike.factors}.}
}
\value{
\code{x} is returned with new assays containing (log-)normalized matrices for endogenous genes and spike-in transcripts.
Size factors are also stored in the \code{\link[SummarizedExperiment]{colData}} for each experiment.
}
\description{
Compute (log-)normalized expression values for endogenous genes and spike-in transcripts after scaling normalization.
All sets of size factors are centered with \code{\link{centerSpikeInFactors}} prior to calling \code{\link{normalizeCounts}} on the relevant assays.
}
\examples{
library(SingleCellExperiment)
sce <- getTestRnaData.se("qc")
sce <- normalizeRnaCountsWithSpikeIns.se(sce, "ERCC")

summary(sce$sizeFactor)
assayNames(sce)

summary(altExp(sce, "ERCC")$sizeFactor)
assayNames(altExp(sce, "ERCC"))

}
\seealso{
\code{\link{normalizeRnaCounts.se}}, for simpler normalization when spike-ins are not present.
}
\author{
Aaron Lun
}
