% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_quickCrisprQc.R
\name{quickCrisprQc.se}
\alias{quickCrisprQc.se}
\alias{formatComputeCrisprQcMetricsResult}
\title{Quick quality control for CRISPR data in a SummarizedExperiment}
\usage{
quickCrisprQc.se(
  x,
  num.threads = 1,
  thresholds = NULL,
  block = NULL,
  more.suggest.args = list(),
  assay.type = "counts",
  output.prefix = NULL,
  meta.name = "qc",
  filter.cells = FALSE
)

formatComputeCrisprQcMetricsResult(compute.res)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object or one of its subclasses.
Rows correspond to CRISPR guides and columns correspond to cells.}

\item{num.threads}{Number of threads, to pass to \code{\link{computeCrisprQcMetrics}}.}

\item{thresholds}{List containing pre-defined thresholds for each QC metric,
see the return value of \code{\link{suggestRnaQcThresholds}} for the expected format.}

\item{block}{Block assignment for each cell, to pass to \code{\link{suggestCrisprQcThresholds}} and \code{\link{filterCrisprQcMetrics}}.}

\item{more.suggest.args}{Named list of additional arguments to pass to \code{\link{suggestCrisprQcThresholds}}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} containing the CRISPR count matrix.}

\item{output.prefix}{String containing a prefix to add to the names of the \code{\link[SummarizedExperiment]{colData}} columns containing the output statistics.}

\item{meta.name}{String containing the name of the \code{\link[S4Vectors]{metadata}} entry containing additional outputs like the filtering thresholds.
If \code{NULL}, additional outputs are not reported.}

\item{filter.cells}{Boolean indicating whether to filter the output to retain only high-quality cells.}

\item{compute.res}{\link[S4Vectors]{DataFrame} returned by \code{\link[scrapper]{computeCrisprQcMetrics}}.}
}
\value{
For \code{quickCrisprQc.se}, a copy of \code{x} is returned with additional columns added to its \code{\link[SummarizedExperiment]{colData}}.
Each column contains per-cell values for one of the QC metrics, see \code{\link[scrapper]{computeCrisprQcMetrics}} for details.
The suggested thresholds are stored as a list in \code{\link[S4Vectors]{metadata}}.
The \code{colData} also contains a \code{keep} column, specifying which cells are to be retained.

If \code{filter.cells=TRUE}, the output object is filtered so that it only contains columns for which \code{keep} is true.
By default, no filtering is performed so the number and order of the columns in the output is the same as \code{x}.

For \code{formatComputeCrisprQcMetricsResult}, a \link[S4Vectors]{DataFrame} is returned with the per-cell QC metrics.
}
\description{
Quickly compute quality control (QC) metrics, thresholds and filters from CRISPR data in a \link[SummarizedExperiment]{SummarizedExperiment}.
This calls \code{\link{computeCrisprQcMetrics}} on an assay in a SummarizedExperiment,
followed by \code{\link{suggestCrisprQcThresholds}} and \code{\link{filterCrisprQcMetrics}} to identify high-quality cells.
}
\examples{
library(SingleCellExperiment)
sce <- altExp(getTestCrisprData.se(), "CRISPR Guide Capture")
sce <- quickCrisprQc.se(sce)
colData(sce)[,c("sum", "detected", "max.value", "max.index")]
metadata(sce)$qc$thresholds
summary(sce$keep)

}
\author{
Aaron Lun
}
