% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_quality_control.R
\name{rna_quality_control_defaults}
\alias{rna_quality_control_defaults}
\alias{computeRnaQcMetricsDefaults}
\alias{suggestRnaQcThresholdsDefaults}
\title{Default parameters for RNA quality control}
\usage{
computeRnaQcMetricsDefaults()

suggestRnaQcThresholdsDefaults()
}
\value{
Named list containing default values for the various function arguments.
}
\description{
Default parameters for RNA quality control
}
\examples{
computeRnaQcMetricsDefaults()
suggestRnaQcThresholdsDefaults()
}
\author{
Aaron Lun
}
