% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPca.R
\name{runPcaDefaults}
\alias{runPcaDefaults}
\title{Default parameters for \code{\link{runPca}}}
\usage{
runPcaDefaults(block = NULL, subset = NULL)
}
\arguments{
\item{block}{See the argument of the same name in \code{\link{runPca}}.}

\item{subset}{See the argument of the same name in \code{\link{runPca}}.}
}
\value{
Named list containing default values for various function arguments.
These defaults may change depending on the values for \code{subset} and \code{block}.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{runPca}} function signature.
}
\examples{
runPcaDefaults()
}
\author{
Aaron Lun
}
