% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTsne.R
\name{runTsneDefaults}
\alias{runTsneDefaults}
\title{Default parameters for \code{\link{runTsne}}}
\usage{
runTsneDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{runTsne}} function signature.
}
\examples{
runTsneDefaults()
}
\author{
Aaron Lun
}
