% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmap.R
\name{runUmapDefaults}
\alias{runUmapDefaults}
\title{Default parameters for \code{\link{runUmap}}}
\usage{
runUmapDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{runUmap}} function signature.
}
\examples{
runUmapDefaults()
}
\author{
Aaron Lun
}
