% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSizeFactors.R
\name{sanitizeSizeFactors}
\alias{sanitizeSizeFactors}
\title{Sanitize size factors}
\usage{
sanitizeSizeFactors(
  size.factors,
  replace.zero = NULL,
  replace.negative = NULL,
  replace.infinite = NULL,
  replace.nan = NULL,
  handle.zero = "sanitize",
  handle.negative = "sanitize",
  handle.infinite = "sanitize",
  handle.nan = "sanitize"
)
}
\arguments{
\item{size.factors}{Numeric vector of size factors across cells.}

\item{replace.zero}{Deprecated, use \code{handle.zero} instead.}

\item{replace.negative}{Deprecated, use \code{handle.negative} instead.}

\item{replace.infinite}{Deprecated, use \code{handle.infinite} instead.}

\item{replace.nan}{Deprecated, use \code{handle.nan} instead.}

\item{handle.zero}{String specifying how to handle replace size factors of zero.
If \code{"sanitize"}, zero size factors are replaced with the lowest positive factor in \code{size.factors}.
This ensures that the normalized values will be large to reflect the extremity of the scaling, but still finite for sensible downstream processing. 
If \code{"error"}, an error is raised upon encountering a size factor of zero.
If \code{"ignore"}, no action is taken.

If \code{NULL}, the default value in \code{\link{sanitizeSizeFactorsDefaults}} is used.}

\item{handle.negative}{String specifying how to handle infinite size factors.
If \code{"sanitize"}, infinite size factors are replaced with the largest positive factor in \code{size.factors}.
This ensures that any normalized values will be, at least, finite; the choice of a relatively large replacement value reflects the extremity of the scaling.
If \code{"error"}, an error is raised upon encountering an infinite size factor. 
If \code{"ignore"}, no action is taken.

If \code{NULL}, the default value in \code{\link{sanitizeSizeFactorsDefaults}} is used.}

\item{handle.nan}{String specifying how to handle NaN size factors.
If \code{"sanitize"}, NaN size factors are replaced with unity, e.g., scaling normalization is a no-op.
If \code{"error"}, an error is raised upon encountering an NaN size factor. 
If \code{"ignore"}, no action is taken.

If \code{NULL}, the default value in \code{\link{sanitizeSizeFactorsDefaults}} is used.}
}
\value{
Numeric vector of length equal to \code{size.factors}, containing the sanitized size factors.
}
\description{
Replace invalid size factors, i.e., zero, negative, infinite or NaN values.
Such size factors can occasionally arise if, e.g., insufficient quality control was performed upstream.
Removing them ensures that the normalized values from \code{\link{normalizeCounts}} remain finite for sensible downstream processing.
}
\examples{
sf <- 2^rnorm(100)
sf[1] <- 0
sf[2] <- -1
sf[3] <- Inf
sf[4] <- NaN
sanitizeSizeFactors(sf)

}
\seealso{
The \code{sanitize_size_factors} function in \url{https://libscran.github.io/scran_norm/}.
}
\author{
Aaron Lun
}
