% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSizeFactors.R
\name{sanitizeSizeFactorsDefaults}
\alias{sanitizeSizeFactorsDefaults}
\title{Default parameters for \code{\link{sanitizeSizeFactors}}}
\usage{
sanitizeSizeFactorsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{sanitizeSizeFactors}} function signature.
}
\examples{
sanitizeSizeFactorsDefaults()
}
\author{
Aaron Lun
}
