% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleByNeighbors.R
\name{scaleByNeighbors}
\alias{scaleByNeighbors}
\title{Scale and combine multiple embeddings}
\usage{
scaleByNeighbors(
  x,
  num.neighbors = 20,
  num.threads = 1,
  weights = NULL,
  BNPARAM = AnnoyParam()
)
}
\arguments{
\item{x}{List of numeric matrices of principal components or other embeddings, one for each modality.
For each entry, rows are dimensions and columns are cells.
All entries should have the same number of columns but may have different numbers of rows.}

\item{num.neighbors}{Integer scalar specifying the number of neighbors to use to define the scaling factor.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}

\item{weights}{Numeric vector of length equal to that of \code{x}, specifying the weights to apply to each modality.
Each value represents a multiplier of the within-population variance of its modality, i.e., larger values increase the contribution of that modality in the combined output matrix.
\code{NULL} is equivalent to an all-1 vector, i.e., all modalities are scaled to have the same within-population variance.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying how to perform the neighbor search.}
}
\value{
List containing \code{scaling}, a vector of scaling factors to be aplied to each embedding;
and \code{combined}, a numeric matrix formed by scaling each entry of \code{x} and then \code{rbind}ing them together.
}
\description{
Scale multiple embeddings (usually derived from different modalities across the same set of cells) so that their within-population variances are comparable,
and then combine them into a single embedding matrix for combined downstream analysis.
}
\examples{
pcs <- list(
    gene = matrix(rnorm(10000), ncol=200),
    protein = matrix(rnorm(1000, sd=3), ncol=200),
    guide = matrix(rnorm(2000, sd=5), ncol=200)
)

out <- scaleByNeighbors.chan(pcs)
out$scaling
dim(out$combined)
}
\author{
Aaron Lun
}
