% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleByNeighbors.R
\name{scaleByNeighborsDefaults}
\alias{scaleByNeighborsDefaults}
\title{Default parameters for \code{\link{scaleByNeighbors}}}
\usage{
scaleByNeighborsDefaults(block = NULL)
}
\arguments{
\item{block}{See the argument of the same name in \code{\link{runPca}}.}
}
\value{
Named list containing default values for various function arguments.
These values may change depending on whether \code{block} is supplied.
}
\description{
Default parameters for \code{\link{scaleByNeighbors}}
}
\examples{
scaleByNeighborsDefaults()
}
\author{
Aaron Lun
}
