% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGeneSet.R
\name{scoreGeneSetDefaults}
\alias{scoreGeneSetDefaults}
\title{Default parameters for \code{\link{scoreGeneSet}}}
\usage{
scoreGeneSetDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{scoreGeneSet}} function signature.
}
\examples{
scoreGeneSetDefaults()
}
\author{
Aaron Lun
}
