% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMarkers.R
\name{scoreMarkersDefaults}
\alias{scoreMarkersDefaults}
\title{Default parameters for \code{\link{scoreMarkers}}}
\usage{
scoreMarkersDefaults(all.pairwise = FALSE)
}
\arguments{
\item{all.pairwise}{See the argument of the same name in \code{\link{scoreMarkers}}.}
}
\value{
Named list containing default values for various function arguments.
These defaults may change depending on the values for \code{all.pairwise}.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{scoreMarkers}} function signature.
}
\examples{
scoreMarkersDefaults()
}
