% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapper-package.R
\docType{package}
\name{scrapper-package}
\alias{scrapper}
\alias{scrapper-package}
\title{scrapper: Bindings to C++ Libraries for Single-Cell Analysis}
\description{
Implements R bindings to C++ code for analyzing single-cell (expression) data, mostly from various libscran libraries. Each function performs an individual step in the single-cell analysis workflow, ranging from quality control to clustering and marker detection. Additional wrappers are provided for easy construction of end-to-end workflows involving Bioconductor objects like SingleCellExperiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/libscran/scrapper}
  \item Report bugs at \url{https://github.com/libscran/scrapper/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

}
