% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleByNeighbors.R
\name{subsampleByNeighborsDefaults}
\alias{subsampleByNeighborsDefaults}
\title{Default parameters for \code{\link{subsampleByNeighbors}}}
\usage{
subsampleByNeighborsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{subsampleByNeighbors}} function signature.
}
\examples{
subsampleByNeighborsDefaults()
}
\author{
Aaron Lun
}
