% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleByPartition.R
\name{subsampleByPartition}
\alias{subsampleByPartition}
\title{Subsample by partition}
\usage{
subsampleByPartition(partitions, number, seed = 2763, force.non.empty = TRUE)
}
\arguments{
\item{partitions}{Factor, integer vector or character vector of length equal to the number of cells,
containing the assignment of each cell to a partition (e.g., clustering).}

\item{number}{Integer specifying the number of cells to retain in the subsample.
If greater than the length of \code{partitions}, all cells are retained.}

\item{seed}{Integer specifying the seed for the random number generator.}

\item{force.non.empty}{Boolean indicating whether each partition should have at least one cell in the subsample.
If \code{FALSE}, partitions may not be represented if the number of cells is less than the ratio of \code{number} to \code{length(partitions)}.}
}
\value{
Integer vector containing the indices of the cells retained in the subsample.
}
\description{
Subsample from a partitioning of cells, preserving the frequency of cells across partitions.
The subsample is typically used as a smaller representative dataset that still exhibits the same distribution of cells as the full dataset.
}
\examples{
x <- matrix(rnorm(10000), nrow=5)
partitions <- clusterKmeans(x, k=20)
keep <- subsampleByPartition(partitions$clusters, 100)
table(partitions$clusters[keep])
sub.x <- x[,keep] # subsample to be used for expensive downstream steps.

# Rare partitions are retained by default:
grouping <- c(LETTERS[1:10], sample(LETTERS[11:26], 990, replace=TRUE))
keep <- subsampleByPartition(grouping, 100)
table(grouping[keep])
keep <- subsampleByPartition(grouping, 100, force.non.empty=FALSE)
table(grouping[keep])

}
\seealso{
\code{\link{subsampleByNeighbors}}, for an alternative downsampling strategy.
}
\author{
Aaron Lun
}
