% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleByPartition.R
\name{subsampleByPartitionDefaults}
\alias{subsampleByPartitionDefaults}
\title{Default parameters for \code{\link{subsampleByPartition}}}
\usage{
subsampleByPartitionDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{subsampleByPartition}} function signature.
}
\examples{
subsampleByPartitionDefaults()
}
\author{
Aaron Lun
}
