% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEnrichment.R
\name{testEnrichment}
\alias{testEnrichment}
\title{Test for gene set enrichment}
\usage{
testEnrichment(x, sets, universe, log = FALSE, num.threads = 1)
}
\arguments{
\item{x}{Vector of identifiers for the interesting genes.}

\item{sets}{List of vectors of identifiers for the pre-defined gene sets.}

\item{universe}{Vector of identifiers for the universe of genes in the dataset.
It is expected that \code{x} is a subset of \code{universe}.
Alternatively, an integer scalar specifying the size of the universe.}

\item{log}{Logical scalar indicating whether to report the log-transformed p-values.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
Numeric vector of (log-transformed) p-values to test for significant enrichment of \code{x} in each entry of \code{sets}.
}
\description{
Perform a hypergeometric test for enrichment of interesting genes (e.g., markers) in one or more pre-defined gene sets.
}
\examples{
testEnrichment(
    x=LETTERS[1:5], 
    sets=list(
        first=LETTERS[1:10],
        second=LETTERS[1:5 * 2],
        third=LETTERS[10:20]
    ),
    universe=LETTERS
)

}
\seealso{
\code{\link{phyper}} and \url{https://github.com/libscran/phyper/},
which is the basis for the underlying calculation.
}
\author{
Aaron Lun
}
