% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSnnGraph.R
\name{buildSnnGraphDefaults}
\alias{buildSnnGraphDefaults}
\title{Default parameters for \code{\link{buildSnnGraph}}}
\usage{
buildSnnGraphDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{buildSnnGraph}} function signature.
}
\examples{
buildSnnGraphDefaults()
}
\author{
Aaron Lun
}
