% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerSizeFactors.R
\name{centerSizeFactors}
\alias{centerSizeFactors}
\title{Center size factors}
\usage{
centerSizeFactors(size.factors, block = NULL, mode = c("lowest", "per-block"))
}
\arguments{
\item{size.factors}{Numeric vector of size factors across cells.}

\item{block}{Vector or factor of length equal to \code{size.factors}, specifying the block of origin for each cell.
Alternatively \code{NULL}, in which case all cells are assumed to be in the same block.}

\item{mode}{String specifying how to scale size factors across blocks.
\code{"lowest"} will scale all size factors by the the lowest per-block average. 
\code{"per-block"} will center the size factors in each block separately.}
}
\value{
Numeric vector of length equal to \code{size.factors}, containing the centered size factors.
}
\description{
Scale the size factors so they are centered at unity,
which ensures that the scale of the counts are preserved (on average) after normalization.
}
\examples{
centerSizeFactors(runif(100))

centerSizeFactors(runif(100), block=sample(3, 100, replace=TRUE))

}
\seealso{
The \code{center_size_factors} function in \url{https://libscran.github.io/scran_norm},
for the rationale of centering the size factors.
}
\author{
Aaron Lun
}
