% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseHighlyVariableGenes.R
\name{chooseHighlyVariableGenes}
\alias{chooseHighlyVariableGenes}
\title{Choose highly variable genes}
\usage{
chooseHighlyVariableGenes(stats, top = 4000, larger = TRUE, keep.ties = TRUE)
}
\arguments{
\item{stats}{Numeric vector of variances (or a related statistic) across all genes.
Typically the residuals from \code{\link{modelGeneVariances}} are used here.}

\item{top}{Integer specifying the number of top genes to retain.}

\item{larger}{Logical scalar indicating whether larger values of \code{stats} correspond to more variable genes.}

\item{keep.ties}{Logical scalar indicating whether to keep tied values of \code{stats}, even if \code{top} may be exceeded.}
}
\value{
Integer vector containing the indices of genes in \code{stats} that are considered to be highly variable.
}
\description{
Choose highly variable genes based on a variance-related statistic.
}
\examples{
resids <- rexp(1000)
str(chooseHighlyVariableGenes(resids))

}
\seealso{
\url{https://libscran.github.io/scran_variances/}, for the underlying implementation.
}
\author{
Aaron Lun
}
