% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePseudoCount.R
\name{choosePseudoCountDefaults}
\alias{choosePseudoCountDefaults}
\title{Default parameters for \code{\link{choosePseudoCount}}}
\usage{
choosePseudoCountDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{choosePseudoCount}} function signature.
}
\examples{
choosePseudoCountDefaults()
}
\author{
Aaron Lun
}
