% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_chooseRnaHvgsWithSpikeIns.R
\name{chooseRnaHvgsWithSpikeIns.se}
\alias{chooseRnaHvgsWithSpikeIns.se}
\title{Choose highly variable genes based on spike-ins}
\usage{
chooseRnaHvgsWithSpikeIns.se(
  x,
  spike.altexp,
  block = NULL,
  num.threads = 1,
  more.endogenous.var.args = list(),
  more.spike.var.args = list(),
  top = 4000,
  more.choose.args = list(),
  assay.type = "logcounts",
  output.prefix = NULL,
  include.per.block = FALSE
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment]{SingleCellExperiment} or one of its subclasses.
The main experiment should contain count data for endogenous genes, where rows correspond to genes and columns correspond to cells.}

\item{spike.altexp}{Integer or string specifying the name/index of the alternative experiment containing the spike-in data.
The assay to use is determined by \code{assay.type}.

Alternatively, \code{spike.altexp} may be a named integer or character vector of length 1.
The name specifies an alternative experiment while the value is the index/name of the assay to use from that experiment.}

\item{block}{Block assignment for each cell, to pass to \code{\link{modelGeneVariances}}.}

\item{num.threads}{Number of threads, to pass to \code{\link{modelGeneVariances}}.}

\item{more.endogenous.var.args}{Named list of additional arguments to pass to \code{\link{modelGeneVariances}} for endogenous genes.
Note that all trend-fitting arguments will be ignored as the mean-variance trend is obtained from the spike-in transcripts.}

\item{more.spike.var.args}{Named list of additional arguments to pass to \code{\link{modelGeneVariances}} for spike-in transcripts.}

\item{top}{Number of HVGs to choose, to pass to \code{\link{chooseHighlyVariableGenes}}.}

\item{more.choose.args}{Named list of arguments to pass to \code{\link{chooseHighlyVariableGenes}}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} containing the log-normalized expression matrix for the RNA data.}

\item{output.prefix}{String containing a prefix to add to the names of the \code{link[SummarizedExperiment]{rowData}} columns containing the output statistics.}

\item{include.per.block}{Logical scalar indicating whether the per-block statistics should be stored in the output \code{rowData}.
Only relevant if \code{block} is specified.}
}
\value{
\code{x} is returned with the per-gene variance modelling statistics for endogenous genes added to its \code{rowData}.
This is equivalent to the columns returned by \code{\link{modelGeneVariances}}, except that:
\itemize{
\item \code{fitted} now contains the interpolated values from the mean-variance trend fitted to the spike-in transcripts.
This represents the technical component of variation for each gene.
\item \code{residuals} now contains the difference of each gene's variance from the trend.
This represents the biological component of variation for each gene.
}
The \code{hvg} column in the \code{rowData} indicates whether a gene was chosen as a HVG based on the largest \code{residuals}.
If \code{include.per.block=TRUE} and \code{block} is specified, the per-block statistics are stored as a nested DataFrame in the \code{per.block} column.

Per-transcript modelling statistics are also added to the \code{rowData} of the alternative experiment containing the spike-in data.
This has all of the same columns returned by \code{\link{modelGeneVariances}}.
}
\description{
Fit a mean-variance trend to spike-in transcripts, and use this to estimate the technical noise for endogenous genes of similar abundance.
Highly variable genes are then selected based on the biological component of variation.
}
\details{
It is generally expected that normalization has been performed with \code{\link{normalizeRnaCountsWithSpikeIns.se}}.
This ensures that the means are comparable between endogenous genes and spike-in transcripts.
}
\examples{
library(SingleCellExperiment)
sce <- getTestRnaData.se("qc")
sce <- normalizeRnaCountsWithSpikeIns.se(sce, "ERCC")

sce <- chooseRnaHvgsWithSpikeIns.se(sce, "ERCC")
summary(rowData(sce)$hvg)

plot(rowData(sce)$means, rowData(sce)$variances, col=factor(rowData(sce)$hvg))
spike.rd <- rowData(altExp(sce, "ERCC"))
points(spike.rd$means, spike.rd$variances, col="dodgerblue", pch=4)
curve(approxfun(spike.rd$means, spike.rd$fitted)(x), col="dodgerblue", add=TRUE)

}
\seealso{
\code{\link{chooseRnaHvgs.se}}, for a simpler function when spike-ins are not available.
}
\author{
Aaron Lun
}
