% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKmeans.R
\name{clusterKmeans}
\alias{clusterKmeans}
\title{K-means clustering}
\usage{
clusterKmeans(
  x,
  k,
  init.method = c("var-part", "kmeans++", "random"),
  refine.method = c("hartigan-wong", "lloyd"),
  var.part.optimize.partition = TRUE,
  var.part.size.adjustment = 1,
  lloyd.iterations = 100,
  hartigan.wong.iterations = 10,
  hartigan.wong.quick.transfer.iterations = 50,
  hartigan.wong.quit.quick.transfer.failure = FALSE,
  seed = 5489L,
  num.threads = 1
)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{k}{Integer scalar specifying the number of clusters.}

\item{init.method}{String specifying the initialization method:
variance partitioning (\code{"var-part"}), kmeans++ (\code{"kmeans++"}) or random initialization (\code{"random"}).}

\item{refine.method}{String specifying the refinement method:
Lloyd's algorithm (\code{"lloyd"}) or the Hartigan-Wong algorithm (\code{"hartigan-wong"}).}

\item{var.part.optimize.partition}{Logical scalar indicating whether each partition boundary should be optimized to reduce the sum of squares in the child partitions.
Only used if \code{init.method = "var.part"}.}

\item{var.part.size.adjustment}{Numeric scalar between 0 and 1, specifying the adjustment to the cluster size when prioritizing the next cluster to partition.
Setting this to 0 will ignore the cluster size while setting this to 1 will generally favor larger clusters.
Only used if \code{init.method = "var.part"}.}

\item{lloyd.iterations}{Integer scalar specifying the maximmum number of iterations for the Lloyd algorithm.}

\item{hartigan.wong.iterations}{Integer scalar specifying the maximmum number of iterations for the Hartigan-Wong algorithm.}

\item{hartigan.wong.quick.transfer.iterations}{Integer scalar specifying the maximmum number of quick transfer iterations for the Hartigan-Wong algorithm.}

\item{hartigan.wong.quit.quick.transfer.failure}{Logical scalar indicating whether to quit the Hartigan-Wong algorithm upon convergence failure during quick transfer iterations.}

\item{seed}{Integer scalar specifying the seed to use for random or kmeans++ initialization.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
By default, a list is returned containing:
\itemize{
\item \code{clusters}, a factor containing the cluster assignments for each cell.
\item \code{centers}, a numeric matrix with the coordinates of the cluster centroids (dimensions in rows, centers in columns).
\item \code{iterations}, an integer scalar specifying the number of refinement iterations that were performed.
\item \code{status}, an integer scalar specifying the convergence status.
Any non-zero value indicates a convergence failure though the exact meaning depends on the choice of \code{refine.method}.
}
}
\description{
Perform k-means clustering with a variety of different initialization and refinement algorithms.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
clustering <- clusterKmeans(x, k=3)
table(clustering$clusters, iris[,"Species"])

}
\author{
Aaron Lun
}
