% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_quality_control.R
\name{crispr_quality_control}
\alias{crispr_quality_control}
\alias{computeCrisprQcMetrics}
\alias{suggestCrisprQcThresholds}
\alias{filterCrisprQcMetrics}
\title{Quality control for CRISPR count data}
\usage{
computeCrisprQcMetrics(x, num.threads = 1)

suggestCrisprQcThresholds(metrics, block = NULL, num.mads = 3)

filterCrisprQcMetrics(thresholds, metrics, block = NULL)
}
\arguments{
\item{x}{A matrix-like object where rows are CRISPRs and columns are cells.
Values are expected to be counts.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}

\item{metrics}{List with the same structure as produced by \code{computeCrisprQcMetrics}.}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for each cell in \code{metrics}.
Alternatively \code{NULL} if all cells are from the same block.

For \code{filterCrisprQcMetrics}, a blocking factor should be provided if \code{block} was used to construct \code{thresholds}.}

\item{num.mads}{Number of median from the median, to define the threshold for outliers in each metric.}

\item{thresholds}{List with the same structure as produced by \code{suggestCrisprQcThresholds}.}
}
\value{
For \code{computeCrisprQcMetrics}, a list is returned containing:
\itemize{
\item \code{sum}, a numeric vector containing the total CRISPR count for each cell.
\item \code{detected}, an integer vector containing the number of detected guides per cell.
\item \code{max.value}, a numeric vector containing the count for the most abundant guide in cell.
\item \code{max.index}, an integer vector containing the row index of the most abundant guide in cell.
}
Each vector is of length equal to the number of cells.

For \code{suggestCrisprQcThresholds}, a named list is returned.
\itemize{
\item If \code{block=NULL}, the list contains:
\itemize{
\item \code{max.value}, a numeric scalar containing the lower bound on the maximum counts for each blocking level.
}
\item Otherwise, if \code{block} is supplied, the list contains:
\itemize{
\item \code{max.value}, a numeric vector containing the lower bound on the maximum counts for each blocking level.
}
Each vector is of length equal to the number of levels in \code{block} and is named accordingly.
}

For \code{filterCrisprQcMetrics}, a logical vector of length \code{ncol(x)} is returned indicating which cells are of high quality.
}
\description{
Compute per-cell QC metrics from an initialized matrix of CRISPR counts,
and use the metrics to suggest filter thresholds to retain high-quality cells.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(100, 100, 0.1) * 100))

qc <- computeCrisprQcMetrics(x)
str(qc)

filt <- suggestCrisprQcThresholds(qc)
str(filt)

keep <- filterCrisprQcMetrics(filt, qc)
summary(keep)

}
\seealso{
The \code{compute_crispr_qc_metrics}, \code{compute_crispr_qc_filters} and \code{compute_crispr_qc_filters_blocked} functions in \url{https://libscran.github.io/scran_qc/},
for the rationale of QC filtering on CRISPR counts.
}
\author{
Aaron Lun
}
