% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTsne.R
\name{runTsne}
\alias{runTsne}
\alias{tsnePerplexityToNeighbors}
\title{t-stochastic neighbor embedding}
\usage{
runTsne(
  x,
  perplexity = 30,
  num.neighbors = tsnePerplexityToNeighbors(perplexity),
  max.depth = 20,
  leaf.approximation = FALSE,
  max.iterations = 500,
  seed = 42,
  num.threads = 1,
  BNPARAM = AnnoyParam()
)

tsnePerplexityToNeighbors(perplexity)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells,
typically containing a low-dimensional representation from, e.g., \code{\link{runPca}}.

Alternatively, a named list of nearest-neighbor search results.
This should contain \code{index}, an integer matrix where rows are neighbors and columns are cells.
Each column contains 1-based indices for the nearest neighbors of the corresponding cell, ordered by increasing distance.
The number of neighbors should be the same as \code{num.neighbors}, otherwise a warning is raised.

Alternatively, an index constructed by \code{\link{buildIndex}}.}

\item{perplexity}{Numeric scalar specifying the perplexity to use in the t-SNE algorithm.}

\item{num.neighbors}{Integer scalar specifying the number of neighbors, typically derived from \code{perplexity}.}

\item{max.depth}{Integer scalar specifying the maximum depth of the Barnes-Hut quadtree.
Smaller values (7-10) improve speed at the cost of accuracy.}

\item{leaf.approximation}{Logical scalar indicating whether to use the \dQuote{leaf approximation} approach,
which sacrifices some accuracy for greater speed.
Only effective when \code{max.depth} is small enough for multiple cells to be assigned to the same leaf node of the quadtree.}

\item{max.iterations}{Integer scalar specifying the maximum number of iterations to perform.}

\item{seed}{Integer scalar specifying the seed to use for generating the initial coordinates.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the algorithm to use.
Only used if \code{x} is not a list of existing nearest-neighbor search results.}
}
\value{
For \code{runTsne}, a numeric matrix where rows are cells and columns are the two dimensions of the embedding.

For \code{tsnePerplexityToNeighbors}, an integer scalar specifying the number of neighbors to use for a given \code{perplexity}.
}
\description{
Compute t-SNE coordinates to visualize similarities between cells.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
embedding <- runTsne(x)
plot(embedding[,1], embedding[,2], col=iris[,5])

}
\seealso{
\url{https://libscran.github.io/qdtsne/}, for an explanation of the approximations.
}
\author{
Aaron Lun
}
