% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmap.R
\name{runUmap}
\alias{runUmap}
\title{Uniform manifold approxation and projection}
\usage{
runUmap(
  x,
  num.dim = 2,
  num.neighbors = 15,
  num.epochs = NULL,
  min.dist = 0.1,
  seed = 1234567890,
  num.threads = 1,
  parallel.optimization = FALSE,
  BNPARAM = AnnoyParam()
)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells,
typically containing a low-dimensional representation from, e.g., \code{\link{runPca}}.

Alternatively, a named list of nearest-neighbor search results like that returned by \code{\link[BiocNeighbors]{findKNN}}.
This should contain \code{index}, an integer matrix where rows are neighbors and columns are cells;
and \code{distance}, a numeric matrix of the same dimensions containing the distances to each neighbor.
Each column contains 1-based indices for the nearest neighbors of the corresponding cell, ordered by increasing distance.
The number of neighbors should be the same as \code{num.neighbors}, otherwise a warning is raised.

Alternatively, an index constructed by \code{\link[BiocNeighbors]{buildIndex}}.}

\item{num.dim}{Integer scalar specifying the number of dimensions of the output embedding.}

\item{num.neighbors}{Integer scalar specifying the number of neighbors to use in the UMAP algorithm.}

\item{num.epochs}{Integer scalar specifying the number of epochs to perform.
If set to \code{NULL}, an appropriate number of epochs is chosen based on \code{ncol(x)}.}

\item{min.dist}{Numeric scalar specifying the minimum distance between points.}

\item{seed}{Integer scalar specifying the seed to use.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}

\item{parallel.optimization}{Logical scalar specifying whether to parallelize the optimization step.}

\item{BNPARAM}{A \link[BiocNeighbors]{BiocNeighborParam} object specifying the algorithm to use.
Only used if \code{x} is not a prebuilt index or a list of existing nearest-neighbor search results.}
}
\value{
A numeric matrix where rows are cells and columns are the two dimensions of the embedding.
}
\description{
Compute UMAP coordinates to visualize similarities between cells.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
embedding <- runUmap(x)
plot(embedding[,1], embedding[,2], col=iris[,5])

}
\seealso{
\url{https://libscran.github.io/umappp/}, for details on the underlying implementation.
}
\author{
Aaron Lun
}
