% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMarkers.R
\name{scoreMarkers}
\alias{scoreMarkers}
\title{Score marker genes}
\usage{
scoreMarkers(
  x,
  groups,
  block = NULL,
  block.weight.policy = c("variable", "equal", "none"),
  variable.block.weight = c(0, 1000),
  compute.auc = TRUE,
  threshold = 0,
  all.pairwise = FALSE,
  num.threads = 1
)
}
\arguments{
\item{x}{A matrix-like object where rows correspond to genes or genomic features and columns correspond to cells.
It is typically expected to contain log-expression values, e.g., from \code{\link{normalizeCounts}}.}

\item{groups}{A vector specifying the group assignment for each cell in \code{x}.}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for each cell in \code{x}.
Alternatively \code{NULL} if all cells are from the same block.}

\item{block.weight.policy}{String specifying the policy to use for weighting different blocks when computing the average for each statistic
(See documentation at \url{https://libscran.github.io/scran_blocks} for more details.)
Only used if \code{block} is not \code{NULL}.}

\item{variable.block.weight}{Numeric vector of length 2, specifying the parameters for variable block weighting.
The first and second values are used as the lower and upper bounds, respectively, for the variable weight calculation.
(See documentation at \url{https://libscran.github.io/scran_blocks} for more details.)
Only used if \code{block} is not \code{NULL} and \code{block.weight.policy = "variable"}.}

\item{compute.auc}{Logical scalar indicating whether to compute the AUC.
Setting this to \code{FALSE} can improve speed and memory efficiency.}

\item{threshold}{Non-negative numeric scalar specifying the minimum threshold on the differences in means (i.e., the log-fold change, if \code{x} contains log-expression values). 
This is incorporated into the effect sizes for Cohen's d and the AUC.}

\item{all.pairwise}{Logical scalar indicating whether to report the full effects for every pairwise comparison between groups.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
If \code{all.pairwise=FALSE}, a list is returned containing:
\itemize{
\item \code{mean}, a numeric matrix containing the mean expression for each group.
Each row is a gene and each column is a group.
\item \code{detected}, a numeric matrix containing the proportion of detected cells in each group.
Each row is a gene and each column is a group.
\item \code{cohens.d}, a list of data frames where each data frame corresponds to a group.
Each row of each data frame represents a gene, while each column contains a summary of Cohen's d from pairwise comparisons to all other groups.
This includes the \code{min}, \code{mean}, \code{median}, \code{max} and \code{min.rank}.
\item \code{auc}, a list like \code{cohens.d} but containing the summaries of the AUCs from each pairwise comparison.
Omitted if \code{compute.auc=FALSE}.
\item \code{delta.mean}, a list like \code{cohens.d} but containing the summaries of the delta-mean from each pairwise comparison.
\item \code{delta.mean}, a list like \code{cohens.d} but containing the summaries of the delta-detected from each pairwise comparison.
}

If \code{all.pairwise=TRUE}, a list is returned containing:
\itemize{
\item \code{mean}, a numeric matrix containing the mean expression for each group.
Each row is a gene and each column is a group.
\item \code{detected}, a numeric matrix containing the proportion of detected cells in each group.
Each row is a gene and each column is a group.
\item \code{cohens.d}, a 3-dimensional array containing the Cohen's from each pairwise comparison between groups.
The first dimension represents the first group, the second dimension represents the second group, and the final dimension represents the gene;
the entry \code{[i, j, k]} represents Cohen's d for \code{i} minus \code{j} of gene \code{k}.
\item \code{auc}, an array like \code{cohens.d} but containing the AUCs from each pairwise comparison.
Omitted if \code{compute.auc=FALSE}.
\item \code{delta.mean}, an array like \code{cohens.d} but containing the delta-mean from each pairwise comparison.
\item \code{delta.detected}, an array like \code{cohens.d} but containing the delta-detected from each pairwise comparison.
}
}
\description{
Score marker genes for each group using a variety of effect sizes from pairwise comparisons between groups.
This includes Cohen's d, the area under the curve (AUC), the difference in the means (delta-mean) and the difference in the proportion of detected cells (delta-detected).
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
normed <- normalizeCounts(x, size.factors=centerSizeFactors(colSums(x)))

# Making up some groups.
g <- sample(letters[1:4], ncol(x), replace=TRUE)
markers <- scoreMarkers(normed, g)
names(markers)
head(markers$mean)
head(markers$cohens.d[["a"]])

}
\seealso{
\url{https://libscran.github.io/scran_variances}, in particular 
the \code{score_markers_summary} function (for \code{all.pairwise=FALSE}),
the \code{score_markers_pairwise} function (for \code{all.pairwise=TRUE}),
and their blocked equivalents \code{score_markers_summary_blocked} and \code{score_markers_pairwise_blocked} (when \code{block} is not \code{NULL}).
}
