% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeEffects.R
\name{summarizeEffects}
\alias{summarizeEffects}
\title{Summarize pairwise effect sizes for each group}
\usage{
summarizeEffects(effects, num.threads = 1)
}
\arguments{
\item{effects}{A 3-dimensional numeric containing the effect sizes from each pairwise comparison between groups.
The extents of the first two dimensions are equal to the number of groups, while the extent of the final dimension is equal to the number of genes.
The entry \code{[i, j, k]} represents Cohen's d from the comparison of group \code{j} over group \code{i} for gene \code{k}.
See also the output of \code{\link{scoreMarkers}} with \code{all.pairwise=TRUE}.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
List of data frames containing summary statistics for the effect sizes.
Each data frame corresponds to a group, each row corresponds to a gene, and each column contains a single summary.
}
\description{
For each group, summarize the effect sizes for all pairwise comparisons to other groups.
This yields a set of summary statistics that can be used to rank marker genes for each group.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
normed <- normalizeCounts(x, size.factors=centerSizeFactors(colSums(x)))

g <- sample(letters[1:4], ncol(x), replace=TRUE)
effects <- scoreMarkers(normed, g, all.pairwise=TRUE)

summarized <- summarizeEffects(effects$cohens.d)
str(summarized)

}
\seealso{
\url{https://libscran.github.io/scran_markers/}, for more details on the statistics.

\code{\link{scoreMarkers}}, to compute the pairwise effects in the first place.
}
\author{
Aaron Lun
}
