% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeEffects.R
\name{summarizeEffectsDefaults}
\alias{summarizeEffectsDefaults}
\title{Default parameters for \code{\link{summarizeEffectsDefaults}}}
\usage{
summarizeEffectsDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters from the underlying C++ library.
These may be overridden by defaults in the \code{\link{summarizeEffects}} function signature.
}
\examples{
summarizeEffectsDefaults()
}
\author{
Aaron Lun
}
