% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEnrichment.R
\name{testEnrichmentDefaults}
\alias{testEnrichmentDefaults}
\title{Default parameters for \code{\link{testEnrichment}}}
\usage{
testEnrichmentDefaults()
}
\value{
Named list containing default values for various function arguments.
}
\description{
Default parameters for \code{\link{testEnrichment}}
}
\examples{
testEnrichmentDefaults()
}
\author{
Aaron Lun
}
